Find text for boxes
The text for text boxes is stored in a very different place from where the boxes are.
The index of the slide and the index of the box should be used to find this text.

Primitive styling is applied to the text. This should be exanded to apply the style to the different parts of the text.