Split up text in styled stretches.
Text in PPT is split in paragraph styled and character styled stretches. They are interleaved.
In the current patch the problem with interleaving is solved by looping over paragraph streches in an out er loop and over all character stretches in an inner loop. For the overlapping parts of the stretches, the text is printed with associated styling information.