added a number of properties and reordered existing ones
added a number of properties and tried to reorder the existing ones into groups to make the file more readable.

I also disabled a number of limitations in the FibBase structure to support files older than Word8.
Word7 and Word6 to be specific.  I added comments to those places.  The point is to identify password protected files and files that we do not support at the moment before initialisation of wv2.

Patch contributed by Matus Uzak